Improve the file structure.
