+++ setup_directory/agent.py
@@ -1,10 +1,10 @@
-from snake_game import SnakeGame 
+from .snake_game import SnakeGame 
 import numpy as np
-import matplotlib.pyplot as plt
-
+import matplotlib.pyplot as plt 
 import random 
 
 class Agent: 
+    
     def __init__( self ) : 
         """
         1. Initialize the agent. 
@@ -23,59 +23,26 @@ class Agent:
 
         return_list = [] 
         for iteration in range( train_count ) : 
-            self.learn( create_visual_at_train ) 
+            self._learn( create_visual_at_train ) 
             if iteration % test_interval == 0 : 
-                result = self.evaluate( trial_count , create_visual_at_test ) 
+                result = self._evaluate( trial_count , create_visual_at_test ) 
                 return_list.append( result ) 
         return return_list 
-
-    def learn( self , create_visual = False ) : 
+    
+    def _learn( self , create_visual = False ) : 
         """ 
         1. Implement the RL algorithm. 
         """ 
         pass 
 
-    def evaluate( self , trial_count , create_visual = False ) : 
+    def _evaluate( self , trial_count , create_visual = False ) : 
         """ 
         1. Test the current model up to trial_count time. 
         2. Return the average result. 
         """ 
         pass 
 
-    ################################## 
-
-    def get_random_action( self , direction ) : 
-        """
-        1. Return a random action. 
-        2. It the previous action is 1 and the current action is 2, this is a guaranteed death. 
-        3. Try to avoid this, as the current function is just to test that the interface is currently working. 
-        """ 
-        how_to_die = { 1 : 2 , 2 : 1 , 3 : 4 , 4 : 3 } 
-        random_integer = random.randint(1, 4) 
-        while random_integer == how_to_die[ direction ] : 
-            random_integer = random.randint(1, 4) 
-        return random_integer 
-
-    def evaluate_random_policy( self , trial_count , create_visual = False ) : 
-        """ 
-        1. Evaluate a random model. 
-        2. Return the reward for each trial. 
-        """
-        return_list = [] 
-        for i in range( trial_count ) : 
-            self.simulator = SnakeGame() 
-            action = 1 
-            done = False 
-            total_reward = 0 
-            while done == False : 
-                action = self.get_random_action( action ) 
-                raw_state , reward , done = self.simulator.step( action , create_visual ) 
-                real_state = self.change_the_state_representation( raw_state ) 
-                total_reward += reward 
-            return_list.append( total_reward ) 
-        return return_list 
-
-    def change_the_state_representation( self , state ) : 
+    def _change_the_state_representation( self , state ) : 
         """ 
         Turn the raw state representation, to a useful state representation. 
         """ 
@@ -106,12 +73,5 @@ class Agent:
 
         # Return the new state representation. 
         return ( state_matrix , food_x , food_y , head_x , head_y , the_direction , the_score ) 
+    
 
-    def test_if_state_matrix_is_correct( self , the_matrix ) : 
-        plt.imshow( the_matrix , cmap='viridis' , interpolation='none' ) 
-        plt.colorbar() 
-        plt.show() 
-
-    ################################## 
-
-
